separates re-mastering of a minion from upgrading salt version

+++ scripts/highstate.sh
@@ -2,15 +2,24 @@
 set -e # everything must pass
 set -u # no unbound variables
 
-# TODO? enabling this will kill any 'stuck' highstate calls
-# killall salt-call
+arg1=$1 # --dry-run
 
-echo "Executing salt highstate (provisioning)"
-log_file=/var/log/salt/salt-highstate-$(date "+%Y-%m-%dT%H:%M:%S").log
-set -o pipefail
-sudo salt-call --force-color state.highstate -l info --retcode-passthrough | tee "$log_file" || {
-    status=$?
-    echo "Error provisioning, state.highstate returned: ${status}"
-    logger "Salt highstate failure: $log_file on $(hostname)"
-    exit $status
-}
+dry_run=false
+if [ "$arg1" = "--dry-run" ]; then
+    dry_run=true
+fi
+
+if $dry_run; then
+    echo "Executing salt highstate (testing)"
+    sudo salt-call --force-color state.highstate -l info test=True --retcode-passthrough
+else
+    echo "Executing salt highstate"
+    log_file=/var/log/salt/salt-highstate-$(date "+%Y-%m-%dT%H:%M:%S").log
+    set -o pipefail
+    sudo salt-call --force-color state.highstate -l info --retcode-passthrough | tee "$log_file" || {
+        status=$?
+        echo "Error provisioning, state.highstate returned: ${status}"
+        logger "Salt highstate failure: $log_file on $(hostname)"
+        exit $status
+    }
+fi
